[fix]: 小bug修复

+++ alg001_MG2250.py
@@ -53,7 +53,7 @@ class Alg001:
         measdate = []
         if not df.empty:
             idx = (df.spd_ref >= algparas[0]) & (df.spd_ref <= algparas[1])
-            re_iter = com_util.Reg.finditer(idx, algparas[2] * df.num_per_sec)
+            re_iter = com_util.Reg.finditer(idx, int(algparas[2] * df.num_per_sec))
             n1 = int(algparas[3] * df.num_per_sec)
             n2 = int(algparas[4] * df.num_per_sec)
             for i in re_iter:
